Refactor for maintainability
Moved the functions for drawing axes to their own file.
