+++ index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 	<head>
-		<title>My first three.js app</title>
+		<title>ScatterPlot3DPlugin</title>
 		<meta charset='utf-8'/>
 		<style>
 			body { margin: 0; }
